+++ bootstrap/src/Language/Generic/Data.hs
@@ -1,5 +1,6 @@
 module Language.Generic.Data
   ( (:+:) (..)
+  , (:++:) (..)
   , type (|:) (..)
   , type (|:$) (..)
   )
@@ -8,7 +9,8 @@ where
 import Data.Functor.Foldable (Recursive)
 import Data.Functor.Foldable.TH (makeBaseFunctor)
 import Tlang.TH (fixQ)
-import Prettyprinter
+import Prettyprinter ( Pretty(pretty) )
+import Data.Bifunctor.TH (deriveBifunctor)
 
 infixr 3 :+:
 
@@ -28,6 +30,21 @@ instance (Show (f a), Show (g a)) => Show ((f :+: g) a) where
 instance (Pretty (f a), Pretty (g a)) => Pretty ((f :+: g) a) where
   pretty (Inl v) = pretty v
   pretty (Inr v) = pretty v
+
+-- | generic bifunctor sum type
+data (f :++: g) b a
+  = Inll (f b a)
+  | Inrr (g b a)
+  deriving (Functor, Foldable, Traversable, Eq, Ord)
+deriveBifunctor ''(:++:)
+
+instance (Show (f b a), Show (g b a)) => Show ((f :++: g) b a) where
+  show (Inll v) = show v
+  show (Inrr v) = show v
+
+instance (Pretty (f b a), Pretty (g b a)) => Pretty ((f :++: g) b a) where
+  pretty (Inll v) = pretty v
+  pretty (Inrr v) = pretty v
 
 -- | a specialised Cofree and its relevant CofreeF
 infixl 2 |: , :|