+++ bootstrap/test/Test/Inference.hs
@@ -13,7 +13,7 @@ import Text.Megaparsec
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit (testCase, testCaseSteps, assertFailure, assertBool)
 import Prettyprinter
-import Transform.Desugar (addSugarToType)
+import Transform.Desugar (pruneForallType)
 
 buildCorrectCase :: Text -> TestTree
 buildCorrectCase text = testCaseSteps (unpack text) \output -> do
@@ -22,7 +22,7 @@ buildCorrectCase text = testCaseSteps (unpack text) \output -> do
     Left err -> assertFailure $ "Parser Error: " <> errorBundlePretty err
     Right e -> return e
   tcExprToSyntacticType [] 0 ("t", 0) e >>= \case
-    Right (t, _) -> output $ show $ pretty $ addSugarToType t
+    Right (t, _) -> output $ show $ pretty $ pruneForallType t
     Left err -> assertFailure $ "Inference Error: " <> show err
 
 buildWrongCase :: Text -> TestTree
@@ -36,7 +36,7 @@ buildWrongCase text = testCase (unpack text) do
     Left err -> assertBool (show err) True
 
 inferringExpressionWithNoError :: TestTree
-inferringExpressionWithNoError = testGroup "Inferring Expression With Error" $
+inferringExpressionWithNoError = testGroup "Inferring Expression With No Error" $
   buildCorrectCase <$>
   [ "9"
   , "9.0"
@@ -60,7 +60,7 @@ inferringExpressionWithNoError = testGroup "Inferring Expression With Error" $
   ]
 
 inferringExpressionWithError :: TestTree
-inferringExpressionWithError = testGroup "Inferring Expression With No Error" $
+inferringExpressionWithError = testGroup "Inferring Expression With Error" $
   buildWrongCase <$>
   [ "let ( ?a, ?b) = 3 in a"
   , "let ( ?a, ?b) = (3,4,5) in a"